// src/app/api/api/models/usuario-response.ts
/* tslint:disable */
/* eslint-disable */
export interface UsuarioResponse {
  activo?: boolean;
  apellidoMaterno?: string;
  apellidoPaterno?: string;
  correo?: string;
  dni?: string;
  esSistema?: boolean;
  fechaCreacion?: string;
  fechaModificacion?: string;
  nombres?: string;
  resetContrasenia?: boolean;
  rolId?: number;
  telefono?: string;
  ultimoAcceso?: string;
  usuario?: string;
  usuarioId?: number;
}